+++ cmd/sparkwing/commands.go
@@ -55,7 +55,7 @@ var allCommands = []*Command{
 	&cmdSparksUpdate, &cmdSparksAdd, &cmdSparksRemove, &cmdSparksWarmup,
 	&cmdApprove, &cmdDeny, &cmdApprovals, &cmdApprovalsList,
 	&cmdAnnotations, &cmdAnnotationsList, &cmdAnnotationsAdd,
-	&cmdRepos, &cmdReposUpdate,
+	&cmdRepos, &cmdReposList, &cmdReposInfo, &cmdReposUpdate,
 }
 
 // CommandJSON is the wire shape emitted by `sparkwing commands` and